MODFEE-108 Create strategy for amount splitting
